streetvision_cascade/requirements.txt
# StreetVision 6-Model Cascade Requirements
# December 20, 2025 - Latest versions per REALISTIC_DEPLOYMENT_PLAN.md

# Core ML Framework
torch>=2.5.0
torchvision>=0.20.0
torchaudio>=2.5.0

# Transformers & Model Loading
transformers>=4.45.0
accelerate>=1.0.0
safetensors>=0.4.0
huggingface-hub>=0.26.0

# vLLM for VLM Serving (v0.12.0 - Dec 4, 2025)
vllm>=0.6.0

# Quantization (AutoAWQ - Superior to GPTQ per plan)
autoawq>=0.2.0

# TensorRT for GPU Optimization
tensorrt>=10.0.0
nvidia-tensorrt

# Object Detection
ultralytics>=8.3.0  # YOLOv12-X

# SDXL for FREE Synthetic Data
diffusers>=0.32.0

# FiftyOne for Hard-Case Mining (v1.5.2)
fiftyone>=1.0.0

# SAM 3 for Concept-Based Annotation (Dec 18, 2025)
# segment-anything-3  # Install when available

# Image Processing
pillow>=10.0.0
opencv-python>=4.8.0
scikit-image>=0.22.0

# Data & Utilities
numpy>=1.26.0
pandas>=2.0.0
scipy>=1.11.0
tqdm>=4.66.0
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Monitoring (Prometheus + Grafana integration)
prometheus-client>=0.19.0
psutil>=5.9.0

# API & Serving
fastapi>=0.109.0
uvicorn>=0.25.0
httpx>=0.26.0
aiohttp>=3.9.0

# Redis for Query Caching (Month 2+)
redis>=5.0.0

# Bittensor
bittensor>=9.0.0
bittensor-cli>=9.0.0

# Weights & Biases (optional)
wandb>=0.16.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Logging
loguru>=0.7.0
rich>=13.0.0

